replace keycloak userid with username field
