+++ submit-api/tests/unit/resources/test_invitation.py
@@ -29,7 +29,7 @@ CONFIG = get_named_config("testing")
 
 def test_create_invitation_to_existing_account(client, session, jwt):
     """Test creating invitation for existing account."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers, account_project = setup_authenticated_proponent(session, jwt)
@@ -53,7 +53,7 @@ def test_create_invitation_to_existing_account(client, session, jwt):
 
 def test_create_invitation_conflict(client, session, jwt):
     """Test creating invitation when user already exists."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     # Mock the service to return a conflict
@@ -79,7 +79,7 @@ def test_create_invitation_conflict(client, session, jwt):
 
 def test_create_new_account_invitation(client, session, jwt):
     """Test creating invitation for new account."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     project = factory_project_model(name="New Project", proponent_id=5678)
@@ -137,7 +137,7 @@ def test_get_invitation_expired(client, session, jwt):
 
 def test_delete_invitation_by_token(client, session, jwt):
     """Test revoking invitation by token."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers, account_project = setup_authenticated_proponent(session, jwt)
@@ -153,7 +153,7 @@ def test_delete_invitation_by_token(client, session, jwt):
 
 def test_delete_invitation_not_found(client, session, jwt):
     """Test revoking non-existent invitation."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers = factory_auth_header(jwt=jwt, claims=TestJwtClaims.staff_admin_role)
@@ -205,7 +205,7 @@ def test_accept_invitation_invalid_token(client, session):
 
 def test_get_invitation_by_id(client, session, jwt):
     """Test retrieving invitation by ID."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers, account_project = setup_authenticated_proponent(session, jwt)
@@ -224,7 +224,7 @@ def test_get_invitation_by_id(client, session, jwt):
 
 def test_get_invitation_by_id_not_found(client, session, jwt):
     """Test retrieving non-existent invitation by ID."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers = factory_auth_header(jwt=jwt, claims=TestJwtClaims.staff_admin_role)
@@ -235,7 +235,7 @@ def test_get_invitation_by_id_not_found(client, session, jwt):
 
 def test_delete_invitation_by_id(client, session, jwt):
     """Test revoking invitation by ID."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers, account_project = setup_authenticated_proponent(session, jwt)
@@ -251,7 +251,7 @@ def test_delete_invitation_by_id(client, session, jwt):
 
 def test_delete_invitation_by_id_not_found(client, session, jwt):
     """Test revoking non-existent invitation by ID."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers = factory_auth_header(jwt=jwt, claims=TestJwtClaims.staff_admin_role)
@@ -262,7 +262,7 @@ def test_delete_invitation_by_id_not_found(client, session, jwt):
 
 def test_resend_invitation(client, session, jwt):
     """Test resending an invitation."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers, account_project = setup_authenticated_proponent(session, jwt)
@@ -278,7 +278,7 @@ def test_resend_invitation(client, session, jwt):
 
 def test_resend_invitation_not_found(client, session, jwt):
     """Test resending non-existent invitation."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers = factory_auth_header(jwt=jwt, claims=TestJwtClaims.staff_admin_role)
@@ -289,7 +289,7 @@ def test_resend_invitation_not_found(client, session, jwt):
 
 def test_resend_used_invitation(client, session, jwt):
     """Test resending already used invitation."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     headers, account_project = setup_authenticated_proponent(session, jwt)
@@ -328,7 +328,7 @@ def test_delete_invitation_unauthorized(client, session):
 
 def test_create_invitation_with_multiple_projects(client, session, jwt):
     """Test creating invitation with multiple account projects."""
-    auth_guid = TestJwtClaims.staff_admin_role['sub']
+    auth_guid = TestJwtClaims.staff_admin_role['preferred_username']
     factory_user_model(auth_guid=auth_guid)
 
     account = factory_account_model(proponent_id=1234)
